requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
slack_sdk
pyppeteer